Moved imgdata to data/asciidata, reworked vis.py accordingly
